Initial development of the backend code
This is an initial build of the backend code which can communicate via api calls in order to use chatGPT3 capabilities. This code can allow for the behaviour of responses to change based on user choice with parameters such as length of response, how the response should be changed if certain words are used and general themes.